tests/submission/PhilipTaylor/canvas/2d.path.stroke.prune.closed.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.path.stroke.prune.closed</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.path.html">path</a>.<a href="index.2d.path.stroke.html">stroke</a>.<a href="index.2d.path.stroke.prune.html">prune</a>.closed</h1>
<p class="desc">Zero-length line segments from closed paths are removed before stroking</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.path.stroke.prune">2d.path.stroke.prune</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="green-100x50.png">
</p><ul id="d"></ul>
<script>
var t = async_test("Zero-length line segments from closed paths are removed before stroking");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);

ctx.strokeStyle = '#f00';
ctx.lineWidth = 100;
ctx.lineCap = 'round';
ctx.lineJoin = 'round';

ctx.beginPath();
ctx.moveTo(50, 25);
ctx.lineTo(50, 25);
ctx.closePath();
ctx.stroke();

_assertPixel(canvas, 50,25, 0,255,0,255, "50,25", "0,255,0,255");


});
</script>

<div id="log"></div>
</body></html>
